+++ libgapbase/gap_val_file.c
@@ -273,7 +273,7 @@ p_write_keylist_value(FILE *fp, GapValKeyList *keyptr, const char *term_str)
         val_ptr = (gint64 *)keyptr->val_ptr;
         l_str = g_strdup_printf("%s%lld%s %s"
                , keyptr->keyword   /* "(keyword " */
-               , *val_ptr          /* value */
+               , (long long int)*val_ptr          /* value */
                , term_ptr
                , keyptr->comment
                );
