+++ libgapvidapi/gap_vid_api_vidindex.c
@@ -321,7 +321,7 @@ p_debug_print_videoindex(t_GVA_Videoindex *vindex)
                , (int)vindex->ofs_tab[l_idx].seek_nr
                , (int)vindex->ofs_tab[l_idx].frame_length
                , (int)vindex->ofs_tab[l_idx].checksum
-               , vindex->ofs_tab[l_idx].timecode_dts
+               , (long long int)vindex->ofs_tab[l_idx].timecode_dts
                );
     }
   }
